Fix: G3-Task 16.3-Backend galería comisiones finalizadas #359. Añadido restricción de que no se pueda cerrar una comisión si no tiene imagen asociada

+++ Backend/src/main/java/com/HolosINC/Holos/commision/CommisionService.java
@@ -380,7 +380,11 @@ public class CommisionService {
 
         if (commission.getStatus() != StatusCommision.ACCEPTED) {
             throw new IllegalStateException("No puedes cerrar una comisión que aún no has empezado.");
-        }    
+        }
+        
+        if (commission.getImage() == null || commission.getImage().length == 0) {
+            throw new IllegalStateException("No puedes cerrar una comisión que no tiene imagen asociada.");
+        }
 
         chatMessageService.deleteConversationByCommisionId(commissionId);
